net, ip family service: Move tests and cleanup
